Update the Inventory Entites and the Event entities
