--- src/main/java/tn/esprit/EldSync/model/Inventory/Inventory.java
@@ -1,29 +1,0 @@
-package tn.esprit.EldSync.model.Inventory;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDate;
-
-@Entity
-@Table(name = "inventory")
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class Inventory {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer inventoryId;
-
-    @ManyToOne
-    @JoinColumn(name = "itemId", referencedColumnName = "itemId")
-    private Item item;
-
-    private Integer quantity;
-    private Integer unitPrice;
-    private LocalDate expiryDate;
-    private String status;
-}
