--- src/main/java/tn/esprit/EldSync/model/Inventory/Item.java
@@ -1,27 +1,0 @@
-package tn.esprit.EldSync.model.Inventory;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDate;
-
-@Entity
-@Table(name = "items")
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class Item {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer itemId;
-
-    private String name;
-    private String description;
-    private Integer quantity;
-    private Integer unitPrice;
-    private LocalDate expiryDate;
-    private String category;
-}
